requirements.txt
django==4.0.2
djangorestframework==3.13.1
pygments==2.11.2
httpie==3.0.2